deploy/code/requirements.txt
flask
slack
slackclient
google-cloud-storage
google-cloud-secret-manager